% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binst.R
\name{create_bins}
\alias{create_bins}
\title{Creates bins given breaks}
\usage{
create_bins(x, breaks)
}
\arguments{
\item{x}{X is a numeric vector which is to be discretized}

\item{breaks}{Breaks are the breaks for the vector X to be broken at. This excludes endpoints}
}
\value{
A vector same length as X is returned with the numeric discretization
}
\description{
Creates bins given breaks
}
\examples{
create_bins(1:10, c(3, 5))
}

