% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binst.R
\name{create_dtbreaks}
\alias{create_dtbreaks}
\title{Create breaks using mdlp}
\usage{
create_dtbreaks(x, y, control = NULL)
}
\arguments{
\item{x}{X is a numeric vector to be discretized}

\item{y}{Y is the response vector used for calculating metrics for discretization}

\item{control}{Control is used for optional parameters for the method}
}
\value{
A vector containing the breaks
}
\description{
Create breaks using mdlp
}
\examples{
dt_breaks <- create_breaks(iris$Sepal.Length, iris$Species, method="dt")
create_bins(iris$Sepal.Length, dt_breaks)
}
\seealso{
\code{\link{create_breaks}}
}

