% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binst.R
\name{create_earthbreaks}
\alias{create_earthbreaks}
\title{Create breaks using earth (i.e. MARS)}
\usage{
create_earthbreaks(x, y, control = NULL)
}
\arguments{
\item{x}{X is a numeric vector to be discretized}

\item{y}{Y is the response vector used for calculating metrics for discretization}

\item{control}{Control is used for optional parameters for the method}
}
\value{
A vector containing the breaks
}
\description{
Create breaks using earth (i.e. MARS)
}
\examples{
earth_breaks <- create_breaks(x=iris$Sepal.Length, y=iris$Sepal.Width, method="earth")
create_bins(iris$Sepal.Length, earth_breaks)
}
\seealso{
\code{\link{create_breaks}}
}

