% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binst.R
\name{create_jenksbreaks}
\alias{create_jenksbreaks}
\title{Create Jenks breaks}
\usage{
create_jenksbreaks(x, control = NULL)
}
\arguments{
\item{x}{X is a numeric vector to be discretized}

\item{control}{Control is used for optional parameters for the method}
}
\value{
A vector containing the breaks
}
\description{
Create Jenks breaks
}
\examples{
jenks_breaks <- create_jenksbreaks(1:10)
create_bins(1:10, jenks_breaks)
}
\seealso{
\code{\link{create_breaks}}
}

