% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binst.R
\name{create_kmeansbreaks}
\alias{create_kmeansbreaks}
\title{create kmeans breaks.}
\usage{
create_kmeansbreaks(x, control = NULL)
}
\arguments{
\item{x}{X is a numeric vector to be discretized}

\item{control}{Control is used for optional parameters for the method}
}
\value{
A vector containing the breaks
}
\description{
create kmeans breaks.
}

