% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binst.R
\name{create_mdlpbreaks}
\alias{create_mdlpbreaks}
\title{Create breaks using mdlp}
\usage{
create_mdlpbreaks(x, y)
}
\arguments{
\item{x}{X is a numeric vector to be discretized}

\item{y}{Y is the response vector used for calculating metrics for discretization}
}
\value{
A vector containing the breaks
}
\description{
Create breaks using mdlp
}
\examples{
entropy_breaks <- create_breaks(1:10, rep(c(1,2), each = 5), method="entropy")
create_bins(1:10, entropy_breaks)
}
\seealso{
\code{\link{create_breaks}}
}

