% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binst.R
\name{get_control}
\alias{get_control}
\title{gets the default parameters for each method.}
\usage{
get_control(method = "kmeans", control = NULL)
}
\arguments{
\item{method}{Method is the type of discretization approach used}

\item{control}{Control are the controls for the algorithm}
}
\value{
List of default parameters
}
\description{
gets the default parameters for each method.
}

