% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binst.R
\name{create_dtbreaks}
\alias{create_dtbreaks}
\title{create breaks using mdlp}
\usage{
create_dtbreaks(x, y, control = NULL)
}
\arguments{
\item{x}{X is a numeric vector to be discretized}

\item{y}{Y is the response vector used for calculating metrics for discretization}

\item{control}{Control is used for optional parameters for the method}
}
\value{
A vector containing the breaks
based on the undocumented function in partykit package
stolen from source code
}
\description{
create breaks using mdlp
}

