% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binst.R
\name{create_mdlpbreaks}
\alias{create_mdlpbreaks}
\title{create breaks using mdlp}
\usage{
create_mdlpbreaks(x, y)
}
\arguments{
\item{x}{X is a numeric vector to be discretized}

\item{y}{Y is the response vector used for calculating metrics for discretization}
}
\value{
A vector containing the breaks
}
\description{
create breaks using mdlp
}

